mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.38",
    "fastapi>=0.117.1",
    "langchain-cohere>=0.4.6",
    "langchain-pinecone>=0.2.12",
    "mcp>=1.14.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
